[Update] Hibernation ad Task Manager tweaks
DisableHibernation did not work for me, I still had hibernation options in control panel. Proper setup should be imho done by command, which I added to original tweak code. This instantly deletes/creates hiberfil.sys file, too. Also after using command for disabling, the EnableHibernation tweak was not able to reenable it. So I updated both counterparts.

ShowTaskManagerDetails should have some time limit to run (I picked 30 sec.). It happened to me that I was servicing a laptop and the script suddenly stuck on this tweak forever. I couldn't test the update on it as I don't have that particular laptop anymore, but I suspect the loop was waiting for the registry value which was never created. The taskmanager issue was fixed by upgrading Win10 to a new build, but still since then, I can't rely on the loop as there's still potential to stuck in case something goes wrong. So I tried to make this workaround.