Disable Microsoft Office 2019 telemetry
Disable the following scheduled tasks responsible for scanning and uploading
user data to Microsoft.

Office ClickToRun Service Monitor: This task monitors the state of your
Microsoft Office ClickToRunSvc and sends crash and error logs to Microsoft.

OfficeTelemetryAgentLogOn2016: This task initiates Office Telemetry Agent,
which scans and uploads usage and error information for Office solutions when
a user logs on to the computer.

OfficeTelemetryAgentFallBack2016: This task initiates the background task for
Office Telemetry Agent, which scans and uploads usage and error information
for Office solutions.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>